+++ Session-11/Library/Transaction.cs
@@ -15,12 +15,14 @@ namespace Library {
         public Guid CustomerID { get; set; }
         public Guid EmployeeID { get; set; }
         public MethodPayment TypeOfPayment { get; set; }
-        public double Cost { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Cost { get; set; }
+        public decimal TotalPrice { get; set; }
+       
 
         public Transaction() {
             ID = Guid.NewGuid();
             Date = DateTime.Now;
+            
         }
     }
 }
